An immediate commit for the current changes of the Element.js Spec plus the introduction to stubs.
== Stubs ==
At the time of commit all changes/additions should have a Spec attached to the change. This can be time consuming, however. As a compromise a Spec stub is defined as a specification without the actual verification. See the description for the iframe example as an example of a Spec stub.

As soon as someone has time, the stub gets filled with the actual verification code.

git-svn-id: http://svn.mootools.net/trunk@1053 4db308c1-fb21-0410-9919-de62e267375e