- simplified **a lot** the Extends Mutator for Class. Fixes some bugs (thanks RD for reporting), and Its now faster than ever, as it doesnt create a new closure for subclassed methods.
Technical Explanation:

Now when you subclass, a property named _parent_ is simply attached to every "subclassed" method. Then, thanks to the caller property, we can retrieve the right parent based on the method we're on by calling "this.parent". This obviously works regardless of how many levels of subclasses we have. The parent assignment used to be in real time (right before each method was called, in a closure, basically how it was done in the original Base.js, not sure how it is now with Base2), but the caller property allows us to avoid it, and all we need now is simply a property.

git-svn-id: http://svn.mootools.net/trunk@1535 4db308c1-fb21-0410-9919-de62e267375e