Implement Observer-Listeners for Events
API looks like:

  var listener = observer.on('click', fn);
  listener.detach();
  listener.attach();
  listener.setFunction(fn2);

  observer.dispatchEvent('click'[, event, bind]);

If the observer has a `getParent` method, it will support bubbling
through objects.

To stop the bubbling, there is a Event object with a `stop()` method.
This Event object has more methods, like `getTarget()` or `getData()`.

The API and implementation looks a lot like DOM Events. This is done
on purpose, so hopefully it will be possible to use these abstracted
Class Events API for the DOM Events as well.